requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface

faiss-cpu

pypdf
pymupdf

sentence-transformers

rouge-score
sentencepiece

pyyaml